Changes to be committed: 	modified:   package-lock.json 	modified:   package.json 	new file:   src/config/database-connection.js 	modified:   src/config/database.js 	modified:   src/models/Avaliacao.js 	modified:   src/models/Evento.js 	modified:   src/server.js 	modified:   src/services/AvaliacaoService.js 	modified:   src/services/CertificadoService.js 	modified:   src/services/EventoService.js 	modified:   src/services/FuncionarioService.js 	modified:   src/services/LocalService.js 	modified:   src/services/PalestranteService.js 	modified:   src/services/PresencaService.js

+++ src/models/Avaliacao.js
@@ -4,7 +4,7 @@ class Avaliacao extends Model {
   static init(sequelize) {
     super.init({
       nota: { 
-        type: DataTypes.INTEGER, 
+        type: DataTypes.FLOAT, // Alterado de INTEGER para FLOAT
         validate: {
           notEmpty: { msg: "Nota da Avaliação deve ser preenchida!" },
           min: { args: [0], msg: "Nota mínima é 0!" },
